+++ tidy3d/components/tcad/materials/charge.py
@@ -1,73 +1,3 @@
-"""Defines heat material specifications"""
-
-from __future__ import annotations
-
-from abc import ABC
-from typing import Tuple
-
-import pydantic.v1 as pd
-
-from tidy3d.constants import (
-    CONDUCTIVITY,
-    ELECTRON_VOLT,
-    PERMITTIVITY,
-    SPECIFIC_HEAT_CAPACITY,
-    THERMAL_CONDUCTIVITY,
-)
-from tidy3d.base import Tidy3dBaseModel
-from tidy3d.data.data_array import SpatialDataArray
-from tidy3d.heat_charge.charge_settings import (
-    AugerRecombination,
-    BandgapNarrowingModelType,
-    CaugheyThomasMobility,
-    MobilityModelType,
-    RadiativeRecombination,
-    RecombinationModelType,
-    SlotboomNarrowingModel,
-    SRHRecombination,
-)
-from tidy3d.types import Union
-
-
-# Liquid class
-class AbstractHeatChargeSpec(ABC, Tidy3dBaseModel):
-    """Abstract heat material specification."""
-
-
-class FluidSpec(AbstractHeatChargeSpec):
-    """Fluid medium. Heat simulations will not solve for temperature
-    in a structure that has a medium with this 'heat_spec'.
-
-    Example
-    -------
-    >>> solid = FluidSpec()
-    """
-
-
-class SolidSpec(AbstractHeatChargeSpec):
-    """Solid medium for heat simulations.
-
-    Example
-    -------
-    >>> solid = SolidSpec(
-    ...     capacity=2,
-    ...     conductivity=3,
-    ... )
-    """
-
-    capacity: pd.PositiveFloat = pd.Field(
-        title="Heat capacity",
-        description=f"Volumetric heat capacity in unit of {SPECIFIC_HEAT_CAPACITY}.",
-        units=SPECIFIC_HEAT_CAPACITY,
-    )
-
-    conductivity: pd.PositiveFloat = pd.Field(
-        title="Thermal conductivity",
-        description=f"Thermal conductivity of material in units of {THERMAL_CONDUCTIVITY}.",
-        units=THERMAL_CONDUCTIVITY,
-    )
-
-
 class ChargeSpec(AbstractHeatChargeSpec):
     """Abstract class for Charge specifications"""
