+++ tidy3d/components/tcad/monitors/heat.py
@@ -1,0 +1,5 @@
+from tidy3d.components.tcad.monitors.abstract import HeatChargeMonitor
+
+
+class TemperatureMonitor(HeatChargeMonitor):
+    """Temperature monitor."""
